+++ lib/shakapacker/configuration.rb
@@ -12,9 +12,6 @@ class Shakapacker::Configuration
   def initialize(root_path:, config_path:, env:)
     @root_path = root_path
     @env = env
-
-    # For backward compatibility
-    Shakapacker.set_shakapacker_env_variables_for_backward_compatibility
     @config_path = Pathname.new(ENV["SHAKAPACKER_CONFIG"] || config_path)
   end
 
@@ -41,17 +38,6 @@ class Shakapacker::Configuration
 
     return false unless config_path.exist?
     fetch(:shakapacker_precompile)
-  end
-
-  def webpacker_precompile?
-    Shakapacker.puts_deprecation_message(
-      Shakapacker.short_deprecation_message(
-        "webpacker_precompile?",
-        "shakapacker_precompile?"
-      )
-    )
-
-    shakapacker_precompile?
   end
 
   def source_path
@@ -103,17 +89,7 @@ class Shakapacker::Configuration
   end
 
   def fetch(key)
-    return data.fetch(key, defaults[key]) unless key == :webpacker_precompile
-
-    # for backward compatibility
-    Shakapacker.puts_deprecation_message(
-      Shakapacker.short_deprecation_message(
-        "webpacker_precompile",
-        "shakapacker_precompile"
-      )
-    )
-
-    data.fetch(key, defaults[:shakapacker_precompile])
+    data.fetch(key, defaults[key])
   end
 
   def asset_host
@@ -148,13 +124,6 @@ class Shakapacker::Configuration
         YAML.load_file(config_path.to_s)
       end
       symbolized_config = config[env].deep_symbolize_keys
-
-      # For backward compatibility
-      if symbolized_config.key?(:shakapacker_precompile) && !symbolized_config.key?(:webpacker_precompile)
-        symbolized_config[:webpacker_precompile] = symbolized_config[:shakapacker_precompile]
-      elsif !symbolized_config.key?(:shakapacker_precompile) && symbolized_config.key?(:webpacker_precompile)
-        symbolized_config[:shakapacker_precompile] = symbolized_config[:webpacker_precompile]
-      end
 
       return symbolized_config
     rescue Errno::ENOENT => e