+++ package/utils/helpers.js
@@ -8,15 +8,6 @@ const chdirTestApp = () => {
   }
 }
 
-// For backward compatibility
-const chdirWebpackerTestApp = () => {
-  try {
-    return process.chdir('spec/backward_compatibility_specs/webpacker_test_app')
-  } catch (e) {
-    return null
-  }
-}
-
 const chdirCwd = () => process.chdir(process.cwd())
 
 const resetEnv = () => {
@@ -64,22 +55,8 @@ const loaderMatches = (configLoader, loaderToCheck, fn) => {
   return fn()
 }
 
-// For backward compatibility
-// Set SHAKAPACKER_XYZ only if SHAKAPACKER_XYZ is not defined but WEBPACKER_XYZ is.
-const setShakapackerEnvVariablesForBackwardCompatibility = () => {
-  const webpackerEnvVariables = Object.entries(process.env).filter((key) => /^WEBPACKER_/.test(key))
-  webpackerEnvVariables.forEach(([webpackerEnvKey, webpackerEnvValue]) => {
-    const shakapackerKey = webpackerEnvKey.replace(/^WEBPACKER/, 'SHAKAPACKER')
-    if (!process.env[shakapackerKey]) {
-      process.env[shakapackerKey] = webpackerEnvValue
-    }
-  })
-}
-
 module.exports = {
   chdirTestApp,
-  chdirWebpackerTestApp,
-  setShakapackerEnvVariablesForBackwardCompatibility,
   chdirCwd,
   isArray,
   isBoolean,